Created ComparerDates function to verify if the initDate and finishDate provided from the body of the request are valid. Verify in getAllController if exist data to show, that is to say, tasks. Verify too, if exist a task with x id in getByController

+++ controller/controller.task.ts
@@ -6,6 +6,17 @@ import { Project } from "../classes/class.project";
 const task: Task = new Task();
 const project: Project = new Project();
 
+const ComparerDates = (initDate: string, finishDate: string, res: Response) => {
+    const initDateToCompare: Date = new Date(initDate);
+    const finishDateToCompare: Date = new Date(finishDate);
+
+    if(initDateToCompare > finishDateToCompare) {
+        return res.status(500).json({
+            message: 'initDate must be minor or equal than finishDate'
+        })
+    }
+}
+
 export const createController = async (req: Request, res: Response) => {
 
     const { id } = req.user;
@@ -16,11 +27,17 @@ export const createController = async (req: Request, res: Response) => {
 
         let { initDate, finishDate } = body;
 
-        if (<Date>initDate.now() > <Date>finishDate.now()) {
-            return res.status(500).json({
-                message: 'initDate must be minor than finishDate'
-            });
-        }
+        // const initDateToCompare: Date = new Date(initDate);
+        // const finishDateToCompare: Date = new Date(finishDate);
+
+        // if (initDateToCompare > finishDateToCompare) {
+        //     return res.status(500).json({
+        //         message: 'initDate must be minor or equal than finishDate'
+        //     });
+
+        // }
+
+        ComparerDates(initDate, finishDate, res);
 
         if (!req.query.projectId) {
             const data = await task.createTask(body, Number(id));
@@ -29,6 +46,8 @@ export const createController = async (req: Request, res: Response) => {
         }
 
         const ifProjectExist = await project.getProjectById(Number(req.query.projectId));
+
+        console.log(ifProjectExist);
 
         if (!ifProjectExist) {
             return res.status(400).json({
@@ -55,6 +74,10 @@ export const getAllController = async (req: Request, res: Response) => {
 
         const data = await task.getAllTask();
 
+        if (data.length == 0) {
+            return res.status(200).json(MessageSuccess('Don`t exist data to show'));
+        }
+
         res.status(200).json(MessageSuccess(data));
 
     }
@@ -74,6 +97,12 @@ export const getByIdController = async (req: Request, res: Response) => {
 
         const data = await task.getTaskById(Number(id));
 
+        if(!data) {
+            res.status(500).json({
+                message: `The task with the id: ${id} don´t exist`
+            });
+        }
+
         res.status(200).json(MessageSuccess(data));
 
     }
@@ -91,6 +120,18 @@ export const updateController = async (req: Request, res: Response) => {
     const { body } = req;
 
     try {
+
+        let { initDate, finishDate } = body;
+
+        ComparerDates(initDate, finishDate, res);
+
+        const verifyIfTaskExist = await task.getTaskById(Number(id));
+
+        if(!verifyIfTaskExist){
+            return res.status(500).json({
+                message: 'The task that you try update don´t exist'
+            });
+        }
 
         const data = await task.updateTask(Number(id), body);
 
@@ -111,6 +152,16 @@ export const deleteController = async (req: Request, res: Response) => { //TODO:
 
     try {
 
+        const verifyIfTaskExist = await task.getTaskById(Number(id));
+
+        console.log(verifyIfTaskExist);
+
+        if(!verifyIfTaskExist){
+            return res.status(500).json({
+                message: 'The task that you try delete don´t exist'
+            });
+        }
+    
         const data = await task.deleteTask(Number(id));
 
         res.status(200).json(MessageSuccess(data.taskDeleted));
